requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
argilla==1.6.0
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.1.0
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.7.4
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.6.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.157
loguru==0.7.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==3.1.0
packaging==23.1
pandas==1.5.3
pandasai==0.5.5
parso==0.8.3
pdfminer.six==20221105
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
prompt-toolkit==3.0.38
protobuf==3.20.3
pure-eval==0.2.2
pyarrow==12.0.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pypandoc==1.11
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.5.4
requests==2.29.0
rfc3986==1.5.0
rich==13.0.1
safetensors==0.3.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@0d824ab77334ed4ab6250e5e491171afeccfb298
soupsieve==2.4.1
SQLAlchemy==1.4.48
sqlitedict==2.1.0
stack-data==0.6.2
streamlit==1.22.0
tenacity==8.2.2
tiktoken==0.3.3
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
unstructured==0.6.2
urllib3==1.26.15
validators==0.20.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
win32-setctime==1.1.0
wrapt==1.14.1
XlsxWriter==3.1.0
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.15.0
